feat(web-stack): extract SR types/utils to web-core
