+++ @xen-orchestra/web/src/modules/storage-repository/composables/use-sr-connect-modal.composable.ts
@@ -1,9 +1,9 @@
 import { useXoPbdPlugJob } from '@/modules/pbd/jobs/xo-pbd-plug.job.ts'
 import { useGetPbdsInScope } from '@/modules/storage-repository/composables/xo-sr-utils.composable.ts'
 import type { FrontXoSr } from '@/modules/storage-repository/remote-resources/use-xo-sr-collection.ts'
-import type { SrScope } from '@/modules/storage-repository/types/storage-repository.type'
-import { getSrAccessMode } from '@/modules/storage-repository/utils/xo-sr.util.ts'
+import type { SrScope } from '@core/types/storage-repository.type.ts'
 import { useModal } from '@core/packages/modal/use-modal.ts'
+import { getSrAccessMode } from '@core/utils/sr.utils.ts'
 import { toComputed } from '@core/utils/to-computed.util.ts'
 import { computed, type MaybeRefOrGetter } from 'vue'
