+++ @xen-orchestra/web/src/modules/vbd/components/actions/disconnect/VbdDisconnectButton.vue
@@ -17,9 +17,9 @@
 import { useVbdConnectionToggleModal } from '@/modules/vbd/composables/use-vbd-connection-toggle-modal.composable.ts'
 import type { FrontXoVbd } from '@/modules/vbd/remote-resources/use-xo-vbd-collection.ts'
 import type { FrontXoVm } from '@/modules/vm/remote-resources/use-xo-vm-collection.ts'
-import { CONNECTION_ACTION } from '@/shared/constants.ts'
-import UiButton from '@xen-orchestra/web-core/components/ui/button/UiButton.vue'
-import { vTooltip } from '@xen-orchestra/web-core/directives/tooltip.directive.ts'
+import UiButton from '@core/components/ui/button/UiButton.vue'
+import { vTooltip } from '@core/directives/tooltip.directive.ts'
+import { CONNECTION_ACTION } from '@core/types/connection.ts'
 import { useI18n } from 'vue-i18n'
 
 const { vbd, vm } = defineProps<{
