+++ @xen-orchestra/web/src/modules/vdi/components/detail/VdiDetailGeneralInfo.vue
@@ -49,7 +49,7 @@ import { type FrontXoVbd, useXoVbdCollection } from '@/modules/vbd/remote-resour
 import { type FrontXoVdi, useXoVdiCollection } from '@/modules/vdi/remote-resources/use-xo-vdi-collection.ts'
 import type { FrontXoVdiSnapshot } from '@/modules/vdi/remote-resources/use-xo-vdi-snapshot-collection.ts'
 import { getVdiIcon } from '@/modules/vdi/utils/xo-vdi.util.ts'
-import { CONNECTION_STATUS, VDI_PAGE_CONTEXT } from '@/shared/constants.ts'
+import { VDI_PAGE_CONTEXT } from '@/shared/constants.ts'
 import VtsStatus from '@core/components/status/VtsStatus.vue'
 import VtsTabularKeyValueList from '@core/components/tabular-key-value-list/VtsTabularKeyValueList.vue'
 import VtsTabularKeyValueRow from '@core/components/tabular-key-value-row/VtsTabularKeyValueRow.vue'
@@ -58,6 +58,7 @@ import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiLink from '@core/components/ui/link/UiLink.vue'
 import UiTagsList from '@core/components/ui/tag/UiTagsList.vue'
 import UiTitle from '@core/components/ui/title/UiTitle.vue'
+import { CONNECTION_STATUS } from '@core/types/connection.ts'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
