+++ @xen-orchestra/web/src/modules/vif/components/actions/disconnect/VifDisconnectButton.vue
@@ -17,9 +17,9 @@
 import { useVifConnectionToggleModal } from '@/modules/vif/composables/use-vif-connection-toggle-modal.composable.ts'
 import type { FrontXoVif } from '@/modules/vif/remote-resources/use-xo-vif-collection.ts'
 import type { FrontXoVm } from '@/modules/vm/remote-resources/use-xo-vm-collection.ts'
-import { CONNECTION_ACTION } from '@/shared/constants.ts'
-import UiButton from '@xen-orchestra/web-core/components/ui/button/UiButton.vue'
-import { vTooltip } from '@xen-orchestra/web-core/directives/tooltip.directive.ts'
+import UiButton from '@core/components/ui/button/UiButton.vue'
+import { vTooltip } from '@core/directives/tooltip.directive.ts'
+import { CONNECTION_ACTION } from '@core/types/connection.ts'
 import { useI18n } from 'vue-i18n'
 
 const { vif, vm } = defineProps<{
