+++ @xen-orchestra/web/src/modules/vif/components/cards/VifGeneralInfoCard.vue
@@ -46,7 +46,6 @@ import { getPoolNetworkRoute } from '@/modules/network/utils/xo-network.util.ts'
 import { useXoPifCollection } from '@/modules/pif/remote-resources/use-xo-pif-collection.ts'
 import { useXoVifUtils } from '@/modules/vif/composables/xo-vif-utils.composable.ts'
 import type { FrontXoVif } from '@/modules/vif/remote-resources/use-xo-vif-collection.ts'
-import { CONNECTION_STATUS } from '@/shared/constants.ts'
 import VtsCodeSnippet from '@core/components/code-snippet/VtsCodeSnippet.vue'
 import VtsStatus from '@core/components/status/VtsStatus.vue'
 import VtsTabularKeyValueList from '@core/components/tabular-key-value-list/VtsTabularKeyValueList.vue'
@@ -54,6 +53,7 @@ import VtsTabularKeyValueRow from '@core/components/tabular-key-value-row/VtsTab
 import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiLink from '@core/components/ui/link/UiLink.vue'
 import UiTitle from '@core/components/ui/title/UiTitle.vue'
+import { CONNECTION_STATUS } from '@core/types/connection.ts'
 import { VIF_LOCKING_MODE } from '@vates/types'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
