+++ @xen-orchestra/web/src/modules/vif/composables/use-vif-connection-toggle-modal.composable.ts
@@ -2,15 +2,13 @@ import { useXoVifConnectJob } from '@/modules/vif/jobs/xo-vif-connect.job.ts'
 import { useXoVifDisconnectJob } from '@/modules/vif/jobs/xo-vif-disconnect.job.ts'
 import type { FrontXoVif } from '@/modules/vif/remote-resources/use-xo-vif-collection.ts'
 import type { FrontXoVm } from '@/modules/vm/remote-resources/use-xo-vm-collection.ts'
-import { CONNECTION_ACTION } from '@/shared/constants.ts'
 import { useModal } from '@core/packages/modal/use-modal.ts'
+import { CONNECTION_ACTION, type ConnectionAction } from '@core/types/connection.ts'
 import { toComputed } from '@core/utils/to-computed.util.ts'
 import { computed, type MaybeRefOrGetter } from 'vue'
 
-export type VifConnectionAction = (typeof CONNECTION_ACTION)[keyof typeof CONNECTION_ACTION]
-
 export function useVifConnectionToggleModal(
-  rawAction: MaybeRefOrGetter<VifConnectionAction>,
+  rawAction: MaybeRefOrGetter<ConnectionAction>,
   rawVifs: MaybeRefOrGetter<FrontXoVif[]>,
   rawVm: MaybeRefOrGetter<FrontXoVm>
 ) {
